requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==3.13
Flask==1.1.1
pymysql==0.9.3
awscli==1.16.166
numpy==1.16.3
pandas==0.24.2
boto3==1.9.156
botocore==1.12.156
s3transfer==0.2.0
scikit-learn==0.21.3
pytest==5.4.1

